requirements.txt
dask==2025.2.0
pandas==2.2.3
apache-airflow>=2.10.5
Flask_AppBuilder==4.5.2
pyarrow==16.1.0